requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
daily-python==0.7.4
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
MouseInfo==0.1.3
openai==1.29.0
pillow==10.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.7.1
pydantic_core==2.18.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
requests==2.31.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
vapi-python==0.1.9
